Health and Energy Scripts
Tracks "Health" and "Energy".

They come with parameters:

Bar - Must be set to a bar manager to change the values.
Draining - For the energy, if true then energy will decrease at "Drain Speed" (per second)
Dead - For health, a public variable set to true if the players health is set to lets then zero.

These parameters are scripts are meant to be attached to the player. To remove/add health/energy, call the add/removal/set methods in these scripts (this will make sure the bar udpates)