Created Crafting UI
The crafting UI has many components:

CraftingUI class manages everything overall, it sets the recipes into the cool scroll box.

RecipeUI holds the data for the recipe so that when selected on it can send it to the crafting UI (for crafting). It also helps handle the item component tooltip.

The HoverBox is an element that will pop up when hovering over a recipeUI. It gets the components of the recipe from the recipeUI and displays them as Components.

ComponentUI is a small element that helps organize the data from a (,amount pair) in a recipe into a cool UI box.

For both HoverBox and the recipe view most of the work is done by a ui 'grid' component that is organizing it for me, letting it crow as we add elements.