Added in liquid resource crafting
I call it "liquid" but it doesn't actually have to be liquids.

There are a lot of names and files in this one, note that recipies and therefore the inventory had small changes.

Now each recipe can have a "liquid resource' that it requires. this shows up in the crafting components with a colored bar.

Also, there is an extra window (the liquid resource window) that has a "ResourceCrafter" on it. This resource crafter lets you see how much of each resource you have, each with their own colored bar. It tracks the resources.

Currently, you can only add carbon by pressing O or T (through the event tester). Not sure how we want liquid resource to be added.

Note that there were quite a few